# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TarjimHub API — professional platform for Arabic interpreters
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"

    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"

  # GLOSSARY
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - name: languagePair
          in: query
          required: false
          schema:
            type: string
        - name: mode
          in: query
          required: false
          schema:
            type: string
        - name: specialty
          in: query
          required: false
          schema:
            type: string
        - name: urgent
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /jobs/{id}/applications:
    get:
      operationId: getJobApplications
      tags: [jobs]
      summary: Get applications for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /jobs/{id}/save:
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save a job for later
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveJobResult"
...
  /messages/{userId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # DASHBOARD
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, primaryLanguagePair, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        email: { type: string }
        role: { type: string }
        primaryLanguagePair: { type: string }
        dialectSpecialty: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        bioAr: { type: ["string", "null"] }
        certifications: { type: ["string", "null"] }
        sessionsCompleted: { type: integer }
        rating: { type: ["number", "null"] }
        isOnline: { type: boolean }
        createdAt: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        bioAr: { type: ["string", "null"] }
        primaryLanguagePair: { type: string }
        dialectSpecialty: { type: ["string", "null"] }
        certifications: { type: ["string", "null"] }
    FollowResult:
      type: object
      required: [following]
      properties:
        following: { type: boolean }

    # GLOSSARY
...
    JobInput:
      type: object
      required: [title, description, languagePair, mode, rateOffered]
      properties:
        title: { type: string }
        description: { type: string }
        languagePair: { type: string }
        dialectPreference: { type: ["string", "null"] }
        mode: { type: string, enum: [otp, video, inperson] }
        rateOffered: { type: string }
        specialty: { type: ["string", "null"] }
        urgent: { type: boolean }
        remote: { type: boolean }
        location: { type: ["string", "null"] }
    Job:
      type: object
      required: [id, clientId, title, description, languagePair, mode, rateOffered, status, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer }
        clientName: { type: ["string", "null"] }
        title: { type: string }
        description: { type: string }
        languagePair: { type: string }
        dialectPreference: { type: ["string", "null"] }
        mode: { type: string }
        rateOffered: { type: string }
        specialty: { type: ["string", "null"] }
        urgent: { type: boolean }
        remote: { type: boolean }
        location: { type: ["string", "null"] }
        status: { type: string }
        applicationsCount: { type: integer }
        createdAt: { type: string }
    ApplicationInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Application:
      type: object
      required: [id, jobId, interpreterId, message, status, appliedAt]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        interpreterId: { type: integer }
        interpreterName: { type: ["string", "null"] }
        message: { type: string }
        status: { type: string }
        appliedAt: { type: string }
    SaveJobResult:
      type: object
      required: [saved]
      properties:
        saved: { type: boolean }

    # FEED
...
    MessageInput:
      type: object
      required: [receiverId, content]
      properties:
        receiverId: { type: integer }
        content: { type: string }
    Message:
      type: object
      required: [id, senderId, receiverId, content, sentAt, read]
      properties:
        id: { type: integer }
        senderId: { type: integer }
        receiverId: { type: integer }
        senderName: { type: ["string", "null"] }
        content: { type: string }
        sentAt: { type: string }
        read: { type: boolean }
...