# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TarjimHub API — professional platform for Arabic interpreters
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  receiverId: number;
  content: string;
}