# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TarjimHub API — professional platform for Arabic interpreters
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  /** @nullable */
  nameAr?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioAr?: string | null;
  primaryLanguagePair?: string;
  /** @nullable */
  dialectSpecialty?: string | null;
  /** @nullable */
  certifications?: string | null;
}